# swagger.yaml
swagger: '2.0'
info:
  title: CHAT APP API
  description: ChatApp API
  version: "1.0.0"
host: sbx-oduz01-20200305.sbx.sms-digital.cloud
schemes:
  - http
  - https
basePath: /
...
security:
  - api_key: []
produces:
  - application/json
paths:
...
  "/fibo/{num}":
    get:
      description: 'Endpoint that calculates the fibonacci series of a number.'
      tags:
        - health
      produces:
        - application/json
      responses:
        '200':
          description: 'Returns 200 for small numbers.'
        '504':
          description: 'Timeout for large numbers.'
...